+++ src/pyimpspec/analysis/kramers_kronig.py
@@ -18,17 +18,32 @@
 # the LICENSES folder.
 
 from dataclasses import dataclass
-from multiprocessing import Pool, Value, cpu_count
-from typing import Callable, List, Tuple, Optional
+from multiprocessing import (
+    Pool,
+    Value,
+    cpu_count,
+)
+from typing import (
+    Callable,
+    List,
+    Tuple,
+    Optional,
+)
 from traceback import format_exc
-from lmfit import minimize, Parameters
+from lmfit import (
+    Parameters,
+    minimize,
+)
 from lmfit.minimizer import MinimizerResult
 from numpy import (
     abs,
     angle,
     array,
     float64,
+    floating,
     inf,
+    integer,
+    issubdtype,
     log10 as log,
     min,
     max,
@@ -37,7 +52,10 @@ from numpy import (
     sum as array_sum,
     zeros,
 )
-from numpy.linalg import pinv, inv
+from numpy.linalg import (
+    inv,
+    pinv,
+)
 from pyimpspec.analysis.fitting import (
     _from_lmfit,
     _to_lmfit,
@@ -96,13 +114,13 @@ class KramersKronigResult:
         return f"KramersKronigResult (num_RC={self.num_RC}, {hex(id(self))})"
 
     def get_frequency(self, num_per_decade: int = -1) -> ndarray:
-        assert type(num_per_decade) is int
+        assert issubdtype(type(num_per_decade), integer), num_per_decade
         if num_per_decade > 0:
             return _interpolate(self.frequency, num_per_decade)
         return self.frequency
 
     def get_impedance(self, num_per_decade: int = -1) -> ndarray:
-        assert type(num_per_decade) is int
+        assert issubdtype(type(num_per_decade), integer), num_per_decade
         if num_per_decade > 0:
             return self.circuit.impedances(self.get_frequency(num_per_decade))
         return self.impedance
@@ -119,7 +137,7 @@ class KramersKronigResult:
         -------
         Tuple[ndarray, ndarray]
         """
-        assert type(num_per_decade) is int
+        assert issubdtype(type(num_per_decade), integer), num_per_decade
         if num_per_decade > 0:
             Z: ndarray = self.get_impedance(num_per_decade)
             return (
@@ -145,7 +163,7 @@ class KramersKronigResult:
         -------
         Tuple[ndarray, ndarray, ndarray]
         """
-        assert type(num_per_decade) is int
+        assert issubdtype(type(num_per_decade), integer), num_per_decade
         if num_per_decade > 0:
             freq: ndarray = self.get_frequency(num_per_decade)
             Z: ndarray = self.circuit.impedances(freq)
@@ -177,16 +195,16 @@ class KramersKronigResult:
 
 def _calculate_tau(i: int, num_RC: int, tau_min: float64, tau_max: float64) -> float:
     # Calculate time constants according to eq. 12 in Schönleber et al. (2014)
-    assert type(i) is int
-    assert type(num_RC) is int
-    assert type(tau_min) is float64
-    assert type(tau_max) is float64
+    assert issubdtype(type(i), integer), i
+    assert issubdtype(type(num_RC), integer), num_RC
+    assert issubdtype(type(tau_min), floating), tau_min
+    assert issubdtype(type(tau_max), floating), tau_max
     return pow(10, (log(tau_min) + (i - 1) / (num_RC - 1) * log(tau_max / tau_min)))
 
 
 def _generate_time_constants(w: ndarray, num_RC: int) -> ndarray:
-    assert type(w) is ndarray
-    assert type(num_RC) is int
+    assert type(w) is ndarray, w
+    assert issubdtype(type(num_RC), integer), num_RC
     taus: ndarray = zeros(shape=(num_RC,))
     tau_min: float64 = 1 / max(w)
     tau_max: float64 = 1 / min(w)
@@ -200,15 +218,15 @@ def _generate_time_constants(w: ndarray, num_RC: int) -> ndarray:
 
 
 def _generate_weight(Z_exp: ndarray) -> ndarray:
-    assert type(Z_exp) is ndarray
+    assert type(Z_exp) is ndarray, Z_exp
     # See eq. 13 in Boukamp (1995)
     return (Z_exp.real**2 + Z_exp.imag**2) ** -1
 
 
 def _calculate_pseudo_chisqr(Z_exp: ndarray, Z_fit: ndarray, weight: ndarray) -> float:
-    assert type(Z_exp) is ndarray
-    assert type(Z_fit) is ndarray
-    assert type(weight) is ndarray
+    assert type(Z_exp) is ndarray, Z_exp
+    assert type(Z_fit) is ndarray, Z_fit
+    assert type(weight) is ndarray, weight
     # See eq. 14 in Boukamp (1995)
     return float(
         array_sum(
@@ -218,7 +236,7 @@ def _calculate_pseudo_chisqr(Z_exp: ndarray, Z_fit: ndarray, weight: ndarray) ->
 
 
 def _calculate_mu(params: Parameters) -> float:
-    assert type(params) is Parameters
+    assert type(params) is Parameters, params
     # Calculates the mu-value based on the fitted parameters according to eq. 21 in
     # Schönleber et al. (2014)
     R_neg: List[float] = []
@@ -249,9 +267,9 @@ def _calculate_mu(params: Parameters) -> float:
 def _elements_to_parameters(
     elements: ndarray, taus: ndarray, add_capacitance: bool
 ) -> Parameters:
-    assert type(elements) is ndarray
-    assert type(taus) is ndarray
-    assert type(add_capacitance) is bool
+    assert type(elements) is ndarray, elements
+    assert type(taus) is ndarray, taus
+    assert type(add_capacitance) is bool, add_capacitance
     parameters: Parameters = Parameters()
     R0: float
     R0, elements = elements[0], elements[1:]
@@ -275,11 +293,11 @@ def _elements_to_parameters(
 def _complex_residual(
     params: Parameters, circuit: Circuit, freq: ndarray, Z_exp: ndarray, weight: ndarray
 ) -> ndarray:
-    assert type(params) is Parameters
-    assert type(circuit) is Circuit
-    assert type(freq) is ndarray
-    assert type(Z_exp) is ndarray
-    assert type(weight) is ndarray
+    assert type(params) is Parameters, params
+    assert type(circuit) is Circuit, circuit
+    assert type(freq) is ndarray, freq
+    assert type(Z_exp) is ndarray, Z_exp
+    assert type(weight) is ndarray, weight
     circuit.set_parameters(_from_lmfit(params))
     Z_fit: ndarray = circuit.impedances(freq)
     return array(
@@ -309,14 +327,14 @@ def _cnls_test(arguments: tuple) -> Tuple[int, float, Circuit, float]:
         method,
         max_nfev,
     ) = arguments
-    assert type(freq) is ndarray
-    assert type(Z_exp) is ndarray
-    assert type(weight) is ndarray
-    assert type(num_RC) is int
-    assert type(add_capacitance) is bool
-    assert type(add_inductance) is bool
-    assert type(method) is str
-    assert type(max_nfev) is int
+    assert type(freq) is ndarray, freq
+    assert type(Z_exp) is ndarray, Z_exp
+    assert type(weight) is ndarray, weight
+    assert issubdtype(type(num_RC), integer), num_RC
+    assert type(add_capacitance) is bool, add_capacitance
+    assert type(add_inductance) is bool, add_inductance
+    assert type(method) is str, method
+    assert issubdtype(type(max_nfev), integer), max_nfev
     w: ndarray = 2 * pi * freq
     taus: ndarray = _generate_time_constants(w, num_RC)
     circuit: Circuit = _generate_circuit(taus, add_capacitance, add_inductance, True)
@@ -387,15 +405,15 @@ def _cnls_mu_process(args: tuple) -> Tuple[int, float, Optional[Circuit], float]
         method,
         max_nfev,
     ) = args
-    assert type(freq) is ndarray
-    assert type(Z_exp) is ndarray
-    assert type(weight) is ndarray
-    assert type(mu_criterion) is float
-    assert type(num_RC) is int
-    assert type(add_capacitance) is bool
-    assert type(add_inductance) is bool
-    assert type(method) is str
-    assert type(max_nfev) is int
+    assert type(freq) is ndarray, freq
+    assert type(Z_exp) is ndarray, Z_exp
+    assert type(weight) is ndarray, weight
+    assert issubdtype(type(mu_criterion), floating), mu_criterion
+    assert issubdtype(type(num_RC), integer), num_RC
+    assert type(add_capacitance) is bool, add_capacitance
+    assert type(add_inductance) is bool, add_inductance
+    assert type(method) is str, method
+    assert issubdtype(type(max_nfev), integer), max_nfev
 
     def exit_early(params: Parameters, i: int, _, *args, **kwargs):
         if 0 <= pool_optimal_num_RC.value < num_RC:  # type: ignore
@@ -444,11 +462,11 @@ def _cnls_mu_process(args: tuple) -> Tuple[int, float, Optional[Circuit], float]
 def _generate_variable_matrices(
     w: ndarray, num_RC: int, taus: ndarray, add_capacitance: bool, abs_Z_exp: ndarray
 ) -> Tuple[ndarray, ndarray]:
-    assert type(w) is ndarray
-    assert type(num_RC) is int
-    assert type(taus) is ndarray
-    assert type(add_capacitance) is bool
-    assert type(abs_Z_exp) is ndarray
+    assert type(w) is ndarray, w
+    assert issubdtype(type(num_RC), integer), num_RC
+    assert type(taus) is ndarray, taus
+    assert type(add_capacitance) is bool, add_capacitance
+    assert type(abs_Z_exp) is ndarray, abs_Z_exp
     # Generate matrices with the following columns (top to bottom is left to right)
     # - R0, series resistance
     # - Ri, resistance in parallel with taus[i - 1] where 0 < i <= num_RC
@@ -479,10 +497,10 @@ def _generate_variable_matrices(
 def _generate_circuit(
     taus: ndarray, add_capacitance: bool, add_inductance: bool, cnls: bool
 ) -> Circuit:
-    assert type(taus) is ndarray
-    assert type(add_capacitance) is bool
-    assert type(add_inductance) is bool
-    assert type(cnls) is bool
+    assert type(taus) is ndarray, taus
+    assert type(add_capacitance) is bool, add_capacitance
+    assert type(add_inductance) is bool, add_inductance
+    assert type(cnls) is bool, cnls
     cdc: List[str] = ["R{R=1}"]
     t: float
     for t in taus:
@@ -513,15 +531,15 @@ def _real_test(
     add_capacitance: bool,
     circuit: Circuit,
 ):
-    assert type(a_re) is ndarray
-    assert type(a_im) is ndarray
-    assert type(Z_exp) is ndarray
-    assert type(abs_Z_exp) is ndarray
-    assert type(w) is ndarray
-    assert type(freq) is ndarray
-    assert type(taus) is ndarray
-    assert type(add_capacitance) is bool
-    assert type(circuit) is Circuit
+    assert type(a_re) is ndarray, a_re
+    assert type(a_im) is ndarray, a_im
+    assert type(Z_exp) is ndarray, Z_exp
+    assert type(abs_Z_exp) is ndarray, abs_Z_exp
+    assert type(w) is ndarray, w
+    assert type(freq) is ndarray, freq
+    assert type(taus) is ndarray, taus
+    assert type(add_capacitance) is bool, add_capacitance
+    assert type(circuit) is Circuit, circuit
     # Fit using the real part
     elements: ndarray = pinv(a_re).dot(Z_exp.real / abs_Z_exp)
     # Fit using the imaginary part to fix the series inductance (and capacitance)
@@ -555,14 +573,14 @@ def _imaginary_test(
     weight: ndarray,
     circuit: Circuit,
 ):
-    assert type(a_im) is ndarray
-    assert type(Z_exp) is ndarray
-    assert type(abs_Z_exp) is ndarray
-    assert type(freq) is ndarray
-    assert type(taus) is ndarray
-    assert type(add_capacitance) is bool
-    assert type(weight) is ndarray
-    assert type(circuit) is Circuit
+    assert type(a_im) is ndarray, a_im
+    assert type(Z_exp) is ndarray, Z_exp
+    assert type(abs_Z_exp) is ndarray, abs_Z_exp
+    assert type(freq) is ndarray, freq
+    assert type(taus) is ndarray, taus
+    assert type(add_capacitance) is bool, add_capacitance
+    assert type(weight) is ndarray, weight
+    assert type(circuit) is Circuit, circuit
     # Fit using the imaginary part
     elements: ndarray = pinv(a_im).dot(Z_exp.imag / abs_Z_exp)
     # Estimate the series resistance
@@ -585,13 +603,13 @@ def _complex_test(
     add_capacitance: bool,
     circuit: Circuit,
 ):
-    assert type(a_re) is ndarray
-    assert type(a_im) is ndarray
-    assert type(Z_exp) is ndarray
-    assert type(abs_Z_exp) is ndarray
-    assert type(taus) is ndarray
-    assert type(add_capacitance) is bool
-    assert type(circuit) is Circuit
+    assert type(a_re) is ndarray, a_re
+    assert type(a_im) is ndarray, a_im
+    assert type(Z_exp) is ndarray, Z_exp
+    assert type(abs_Z_exp) is ndarray, abs_Z_exp
+    assert type(taus) is ndarray, taus
+    assert type(add_capacitance) is bool, add_capacitance
+    assert type(circuit) is Circuit, circuit
     # Fit using the complex impedance
     x: ndarray = inv(a_re.T.dot(a_re) + a_im.T.dot(a_im))
     y: ndarray = a_re.T.dot(Z_exp.real / abs_Z_exp) + a_im.T.dot(Z_exp.imag / abs_Z_exp)
@@ -609,12 +627,12 @@ def _test_wrapper(arguments: tuple) -> Tuple[int, float, Circuit, float]:
     num_RC: int
     add_capacitance: bool
     test, freq, Z_exp, weight, num_RC, add_capacitance = arguments
-    assert type(test) is str
-    assert type(freq) is ndarray
-    assert type(Z_exp) is ndarray
-    assert type(weight) is ndarray
-    assert type(num_RC) is int
-    assert type(add_capacitance) is bool
+    assert type(test) is str, test
+    assert type(freq) is ndarray, freq
+    assert type(Z_exp) is ndarray, Z_exp
+    assert type(weight) is ndarray, weight
+    assert issubdtype(type(num_RC), integer), num_RC
+    assert type(add_capacitance) is bool, add_capacitance
     abs_Z_exp: ndarray = abs(Z_exp)
     w: ndarray = 2 * pi * freq
     taus: ndarray = _generate_time_constants(w, num_RC)
@@ -720,12 +738,12 @@ def perform_test(
         type(test),
         test,
     )
-    assert type(num_RC) is int, (
+    assert issubdtype(type(num_RC), integer), (
         type(num_RC),
         num_RC,
     )
     assert num_RC <= data.get_num_points(), f"{num_RC=} > {data.get_num_points()=}"
-    assert type(mu_criterion) is float, (
+    assert issubdtype(type(mu_criterion), floating), (
         type(mu_criterion),
         mu_criterion,
     )
@@ -743,11 +761,11 @@ def perform_test(
         type(method),
         method,
     )
-    assert type(max_nfev) is int, (
+    assert issubdtype(type(max_nfev), integer), (
         type(max_nfev),
         max_nfev,
     )
-    assert type(num_procs) is int, (
+    assert issubdtype(type(num_procs), integer), (
         type(num_procs),
         num_procs,
     )
@@ -938,12 +956,12 @@ def perform_exploratory_tests(
         type(num_RCs),
         num_RCs,
     )
-    assert all(map(lambda _: type(_) is int, num_RCs))
+    assert all(map(lambda _: issubdtype(type(_), integer), num_RCs))
     if len(num_RCs) > 0:
         assert (
             max(num_RCs) <= data.get_num_points()
         ), f"{max(num_RCs)=} > {data.get_num_points()=}"
-    assert type(mu_criterion) is float, (
+    assert issubdtype(type(mu_criterion), floating), (
         type(mu_criterion),
         mu_criterion,
     )
@@ -960,11 +978,11 @@ def perform_exploratory_tests(
         type(method),
         method,
     )
-    assert type(max_nfev) is int, (
+    assert issubdtype(type(max_nfev), integer), (
         type(max_nfev),
         max_nfev,
     )
-    assert type(num_procs) is int, (
+    assert issubdtype(type(num_procs), integer), (
         type(num_procs),
         num_procs,
     )
@@ -1074,7 +1092,7 @@ def score_test_results(
         results,
     )
     assert all(map(lambda _: type(_) is KramersKronigResult, results)), results
-    assert type(mu_criterion) is float, (
+    assert issubdtype(type(mu_criterion), floating), (
         type(mu_criterion),
         mu_criterion,
     )