+++ src/pyimpspec/circuit/parallel.py
@@ -17,8 +17,15 @@
 # The licenses of pyimpspec's dependencies and/or sources of portions of code are included in
 # the LICENSES folder.
 
-from typing import List, Union, Tuple
-from .base import Connection, Element
+from typing import (
+    List,
+    Union,
+    Tuple,
+)
+from .base import (
+    Connection,
+    Element,
+)
 
 
 class Parallel(Connection):