+++ src/pyimpspec/circuit/resistor.py
@@ -17,7 +17,10 @@
 # The licenses of pyimpspec's dependencies and/or sources of portions of code are included in
 # the LICENSES folder.
 
-from typing import Dict, List
+from typing import (
+    Dict,
+    List,
+)
 from collections import OrderedDict
 from .base import Element
 from numpy import inf