+++ tests/test_fitting.py
@@ -37,7 +37,7 @@ from typing import List
 
 
 seed(42)
-data: DataSet = DataSet.from_dict(
+DATA: DataSet = DataSet.from_dict(
     {
         "frequency": [
             10000,
@@ -135,7 +135,7 @@ data: DataSet = DataSet.from_dict(
     }
 )
 sd: float = 0.005
-data.subtract_impedance(
+DATA.subtract_impedance(
     -array(
         list(
             map(
@@ -143,7 +143,7 @@ data.subtract_impedance(
                     abs(_) * normal(0, sd, 1),
                     abs(_) * normal(0, sd, 1),
                 ),
-                data.get_impedance(),
+                DATA.get_impedance(),
             )
         )
     )
@@ -153,26 +153,26 @@ data.subtract_impedance(
 class TestKramersKronig(TestCase):
     def test_01_test(self):
         # complex
-        test: KramersKronigResult = perform_test(data)
+        test: KramersKronigResult = perform_test(DATA)
         self.assertTrue(type(test) is KramersKronigResult)
         self.assertEqual(test.num_RC, 2)
         self.assertAlmostEqual(test.mu, 0.8465317822199895)
         self.assertAlmostEqual(test.pseudo_chisqr, 5.170991769675877)
         self.assertEqual(test.circuit.to_string(), "[RKKL]")
         # real
-        test = perform_test(data, test="real")
+        test = perform_test(DATA, test="real")
         self.assertEqual(test.num_RC, 15)
         self.assertAlmostEqual(test.mu, 0.8351972078416112)
         self.assertAlmostEqual(test.pseudo_chisqr, 0.0022502074349343235)
         self.assertEqual(test.circuit.to_string(), "[RKKKKKKKKKKKKKKKL]")
         # imaginary
-        test = perform_test(data, test="imaginary")
+        test = perform_test(DATA, test="imaginary")
         self.assertEqual(test.num_RC, 2)
         self.assertAlmostEqual(test.mu, 0.7634133056991227)
         self.assertAlmostEqual(test.pseudo_chisqr, 5.397326608733074)
         self.assertEqual(test.circuit.to_string(), "[RKKL]")
         # cnls
-        test = perform_test(data, test="cnls", max_nfev=1000)
+        test = perform_test(DATA, test="cnls", max_nfev=1000)
         self.assertEqual(test.num_RC, 14)
         self.assertAlmostEqual(test.mu, 0.8309531383631978)
         self.assertAlmostEqual(test.pseudo_chisqr, 0.0013382040618625823)
@@ -180,92 +180,92 @@ class TestKramersKronig(TestCase):
 
     def test_02_add_capacitance(self):
         # complex
-        test: KramersKronigResult = perform_test(data, add_capacitance=True)
+        test: KramersKronigResult = perform_test(DATA, add_capacitance=True)
         self.assertTrue("C" in test.circuit.to_string())
         # real
-        test = perform_test(data, test="real", add_capacitance=True)
+        test = perform_test(DATA, test="real", add_capacitance=True)
         self.assertTrue("C" in test.circuit.to_string())
         # imaginary
-        test = perform_test(data, test="imaginary", add_capacitance=True)
+        test = perform_test(DATA, test="imaginary", add_capacitance=True)
         self.assertTrue("C" in test.circuit.to_string())
         # cnls
-        test = perform_test(data, test="cnls", add_capacitance=True, max_nfev=1000)
+        test = perform_test(DATA, test="cnls", add_capacitance=True, max_nfev=1000)
         self.assertTrue("C" in test.circuit.to_string())
 
     def test_03_add_inductance(self):
         # complex
-        test: KramersKronigResult = perform_test(data, add_inductance=True)
+        test: KramersKronigResult = perform_test(DATA, add_inductance=True)
         self.assertTrue("L" in test.circuit.to_string())
         # real
-        test = perform_test(data, test="real", add_inductance=True)
+        test = perform_test(DATA, test="real", add_inductance=True)
         self.assertTrue("L" in test.circuit.to_string())
         # imaginary
-        test = perform_test(data, test="imaginary", add_inductance=True)
+        test = perform_test(DATA, test="imaginary", add_inductance=True)
         self.assertTrue("L" in test.circuit.to_string())
         # cnls
-        test = perform_test(data, test="cnls", add_inductance=True, max_nfev=1000)
+        test = perform_test(DATA, test="cnls", add_inductance=True, max_nfev=1000)
         self.assertTrue("L" in test.circuit.to_string())
 
     def test_04_add_capacitance_inductance(self):
         # complex
         test: KramersKronigResult = perform_test(
-            data, add_capacitance=True, add_inductance=True
+            DATA, add_capacitance=True, add_inductance=True
         )
         self.assertTrue("CL" in test.circuit.to_string())
         # real
         test = perform_test(
-            data, test="real", add_capacitance=True, add_inductance=True
+            DATA, test="real", add_capacitance=True, add_inductance=True
         )
         self.assertTrue("CL" in test.circuit.to_string())
         # imaginary
         test = perform_test(
-            data, test="imaginary", add_capacitance=True, add_inductance=True
+            DATA, test="imaginary", add_capacitance=True, add_inductance=True
         )
         self.assertTrue("CL" in test.circuit.to_string())
         # cnls
         test = perform_test(
-            data, test="cnls", add_capacitance=True, add_inductance=True, max_nfev=1000
+            DATA, test="cnls", add_capacitance=True, add_inductance=True, max_nfev=1000
         )
         self.assertTrue("CL" in test.circuit.to_string())
 
     def test_05_num_RC(self):
         num_RC: int = 10
         # complex
-        test: KramersKronigResult = perform_test(data, num_RC=num_RC)
+        test: KramersKronigResult = perform_test(DATA, num_RC=num_RC)
         self.assertEqual(test.num_RC, num_RC)
         self.assertEqual(test.circuit.to_string().count("K"), num_RC)
         # real
-        test = perform_test(data, test="real", num_RC=num_RC)
+        test = perform_test(DATA, test="real", num_RC=num_RC)
         self.assertEqual(test.num_RC, num_RC)
         self.assertEqual(test.circuit.to_string().count("K"), num_RC)
         # imaginary
-        test = perform_test(data, test="imaginary", num_RC=num_RC)
+        test = perform_test(DATA, test="imaginary", num_RC=num_RC)
         self.assertEqual(test.num_RC, num_RC)
         self.assertEqual(test.circuit.to_string().count("K"), num_RC)
         # cnls
-        test = perform_test(data, test="cnls", num_RC=num_RC, max_nfev=1000)
+        test = perform_test(DATA, test="cnls", num_RC=num_RC, max_nfev=1000)
         self.assertEqual(test.num_RC, num_RC)
         self.assertEqual(test.circuit.to_string().count("K"), num_RC)
 
     def test_06_mu_criterion(self):
         mu_criterion: float = 0.74
         # complex
-        test: KramersKronigResult = perform_test(data, mu_criterion=mu_criterion)
+        test: KramersKronigResult = perform_test(DATA, mu_criterion=mu_criterion)
         self.assertTrue(test.mu <= mu_criterion)
         # real
-        test = perform_test(data, test="real", mu_criterion=mu_criterion)
+        test = perform_test(DATA, test="real", mu_criterion=mu_criterion)
         self.assertTrue(test.mu <= mu_criterion)
         # imaginary
-        test = perform_test(data, test="imaginary", mu_criterion=mu_criterion)
+        test = perform_test(DATA, test="imaginary", mu_criterion=mu_criterion)
         self.assertTrue(test.mu <= mu_criterion)
         # cnls
-        test = perform_test(data, test="cnls", mu_criterion=mu_criterion, max_nfev=1000)
+        test = perform_test(DATA, test="cnls", mu_criterion=mu_criterion, max_nfev=1000)
         self.assertTrue(test.mu <= mu_criterion)
 
     def test_07_threading(self):
         # complex
-        test_single_thread: KramersKronigResult = perform_test(data, num_procs=1)
-        test_multithreaded: KramersKronigResult = perform_test(data, num_procs=2)
+        test_single_thread: KramersKronigResult = perform_test(DATA, num_procs=1)
+        test_multithreaded: KramersKronigResult = perform_test(DATA, num_procs=2)
         self.assertEqual(test_single_thread.num_RC, 2)
         self.assertAlmostEqual(test_single_thread.mu, 0.8465317822199895)
         self.assertAlmostEqual(test_single_thread.pseudo_chisqr, 5.170991769675877)
@@ -276,10 +276,10 @@ class TestKramersKronig(TestCase):
         self.assertEqual(test_multithreaded.circuit.to_string(), "[RKKL]")
         # cnls
         test_single_thread: KramersKronigResult = perform_test(
-            data, test="cnls", num_procs=1, max_nfev=1000
+            DATA, test="cnls", num_procs=1, max_nfev=1000
         )
         test_multithreaded: KramersKronigResult = perform_test(
-            data, test="cnls", num_procs=2, max_nfev=1000
+            DATA, test="cnls", num_procs=2, max_nfev=1000
         )
         self.assertEqual(test_single_thread.num_RC, 14)
         self.assertAlmostEqual(test_single_thread.mu, 0.8309531383631978)
@@ -295,7 +295,7 @@ class TestKramersKronig(TestCase):
         tests: List[KramersKronigResult]
         test: KramersKronigResult
         # complex
-        tests = perform_exploratory_tests(data, mu_criterion=mu_criterion)
+        tests = perform_exploratory_tests(DATA, mu_criterion=mu_criterion)
         self.assertTrue(
             type(tests) is list
             and all(map(lambda _: type(_) is KramersKronigResult, tests))
@@ -303,17 +303,17 @@ class TestKramersKronig(TestCase):
         test = score_test_results(tests, mu_criterion)[0][1]
         self.assertTrue(test.mu <= mu_criterion)
         # real
-        tests = perform_exploratory_tests(data, test="real")
+        tests = perform_exploratory_tests(DATA, test="real")
         # imaginary
-        tests = perform_exploratory_tests(data, test="imaginary")
+        tests = perform_exploratory_tests(DATA, test="imaginary")
         # cnls
-        tests = perform_exploratory_tests(data, test="cnls", max_nfev=1000)
+        tests = perform_exploratory_tests(DATA, test="cnls", max_nfev=1000)
 
 
 class TestFitting(TestCase):
     def test_01_default(self):
         circuit: Circuit = string_to_circuit("R(RC)(RW)")
-        fit: FittingResult = fit_circuit_to_data(circuit, data)
+        fit: FittingResult = fit_circuit_to_data(circuit, DATA)
         self.assertEqual(
             fit.circuit.to_string(0),
             "[R{R=1E+02/0E+00}(R{R=2E+02/0E+00}C{C=8E-07/0E+00/1E+03})(R{R=5E+02/0E+00}W{Y=4E-04/0E+00})]",
@@ -328,14 +328,92 @@ class TestFitting(TestCase):
         self.assertAlmostEqual(param.value, 5.03E2, delta=2E0)
         param: FittedParameter = fit.parameters["W_4"]["Y"]
         self.assertAlmostEqual(param.value, 4.00E-4, delta=1E-5)
+        # Markdown table
+        lines: List[str] = fit.to_dataframe().to_markdown().split("\n")
+        self.assertEqual(len(lines), 7)
+        line: str = lines.pop(0)
+        self.assertTrue(
+            0
+            < line.index("Element")
+            < line.index("Parameter")
+            < line.index("Value")
+            < line.index("Std. err. (%)")
+            < line.index("Fixed")
+        )
+        lines.pop(0)
+        i: int = 0
+        while lines:
+            line = lines.pop(0)
+            columns: List[str] = list(
+                filter(lambda _: _ != "", map(str.strip, line.split("|")))
+            )
+            self.assertEqual(len(columns), 6)
+            self.assertEqual(int(columns[0]), i)
+            self.assertTrue(columns[1].endswith(f"_{i}"))
+            self.assertTrue(columns[1] in fit.parameters)
+            self.assertAlmostEqual(
+                float(columns[3]),
+                fit.parameters[columns[1]][columns[2]].value,
+                delta=0.1 * fit.parameters[columns[1]][columns[2]].value,
+            )
+            self.assertEqual(
+                columns[5],
+                "Yes"
+                if fit.parameters[columns[1]][columns[2]].fixed is True
+                else "No",
+            )
+            i += 1
+        # LaTeX table
+        lines = fit.to_dataframe().to_latex().split("\n")
+        self.assertEqual(lines.pop(0), r"\begin{tabular}{lllrrl}")
+        self.assertEqual(lines.pop(0), r"\toprule")
+        line = lines.pop(0)
+        self.assertTrue(
+            0
+            < line.index("Element")
+            < line.index("Parameter")
+            < line.index("Value")
+            < line.index(r"Std. err. (\%)")
+            < line.index("Fixed")
+        )
+        self.assertEqual(lines.pop(0), r"\midrule")
+        self.assertEqual(lines.pop(), "")
+        self.assertEqual(lines.pop(), r"\end{tabular}")
+        self.assertEqual(lines.pop(), r"\bottomrule")
+        i = 0
+        while lines:
+            line = lines.pop(0).replace(r"\\", "").strip()
+            if line == "":
+                continue
+            columns: List[str] = list(
+                filter(lambda _: _ != "", map(str.strip, line.split("&")))
+            )
+            self.assertEqual(len(columns), 6)
+            self.assertEqual(int(columns[0]), i)
+            self.assertTrue(columns[1].endswith(f"_{i}"))
+            self.assertTrue(columns[1].replace(r"\_", "_") in fit.parameters)
+            self.assertAlmostEqual(
+                float(columns[3]),
+                fit.parameters[columns[1].replace(r"\_", "_")][columns[2]].value,
+                delta=0.1
+                * fit.parameters[columns[1].replace(r"\_", "_")][columns[2]].value,
+            )
+            self.assertEqual(
+                columns[5],
+                "Yes"
+                if fit.parameters[columns[1].replace(r"\_", "_")][columns[2]].fixed
+                is True
+                else "No",
+            )
+            i += 1
 
     def test_02_threading(self):
         circuit: Circuit = string_to_circuit("R(RC)(RW)")
         fit_single_thread: FittingResult = fit_circuit_to_data(
-            circuit, data, max_nfev=100, num_procs=1
+            circuit, DATA, max_nfev=100, num_procs=1
         )
         fit_multithreaded: FittingResult = fit_circuit_to_data(
-            circuit, data, max_nfev=100, num_procs=2
+            circuit, DATA, max_nfev=100, num_procs=2
         )
         self.assertEqual(
             fit_single_thread.circuit.to_string(2),